moved unused function to shared.helper
