+++ mypy.ini
@@ -3,12 +3,12 @@
     disallow_untyped_defs = True
     check_untyped_defs = True
 
-[mypy-molbe.*]
+[mypy-quemb.molbe.*]
     disallow_untyped_defs = False
     check_untyped_defs = False
 
 
-[mypy-kbe.*]
+[mypy-quemb.kbe.*]
     disallow_untyped_defs = False
     check_untyped_defs = False
