+++ src/quemb/__init__.py
@@ -1,3 +1,3 @@
-from quemb import shared, kbe, molbe
+from quemb import kbe, molbe, shared
 
 __all__ = ["molbe", "kbe", "shared"]
