+++ src/quemb/kbe/solver.py
@@ -3,7 +3,7 @@
 import numpy
 import scipy.linalg
 
-from quemb.molbe.helper import unused
+from quemb.shared.helper import unused
 
 
 def schmidt_decomp_svd(rdm, Frag_sites):
