+++ src/quemb/molbe/be_parallel.py
@@ -15,7 +15,6 @@ from quemb.molbe.helper import (
     get_frag_energy,
     get_frag_energy_u,
     get_scfObj,
-    unused,
 )
 from quemb.molbe.solver import (
     make_rdm1_ccsd_t1,
@@ -25,6 +24,7 @@ from quemb.molbe.solver import (
     solve_mp2,
     solve_uccsd,
 )
+from quemb.shared.helper import unused
 
 
 def run_solver(
