+++ src/quemb/molbe/lo.py
@@ -11,7 +11,8 @@ from quemb.molbe.external.lo_helper import (
     get_aoind_by_atom,
     reorder_by_atom_,
 )
-from quemb.molbe.helper import ncore_, unused
+from quemb.molbe.helper import ncore_
+from quemb.shared.helper import unused
 
 
 def dot_gen(A, B, ovlp):
