+++ src/quemb/shared/helper.py
@@ -1,0 +1,6 @@
+# Author(s): Oskar Weser
+
+
+def unused(*args) -> None:
+    for arg in args:
+        del arg
